+++ init.rb
@@ -11,12 +11,13 @@ require 'redmine'
 
 require File.expand_path('../lib/redmine_expert_metrics/collector', __FILE__)
 require File.expand_path('../lib/redmine_expert_metrics/exposition', __FILE__)
+require File.expand_path('../lib/redmine_expert_metrics/mail_observer', __FILE__)
 
 Redmine::Plugin.register :redmine_expert_metrics do
   name 'Redmine expert Metrics'
   author 'Dennis Buehring'
   description 'Shows who is currently working in Redmine and exposes that (plus a few basic totals) as Prometheus metrics. Works on Redmine 5.1 - 7.x.'
-  version '1.0.0'
+  version '1.1.0'
   url 'https://github.com/expertZentrale/redmine_expert_metrics'
   requires_redmine :version_or_higher => '5.0'
 
@@ -38,3 +39,7 @@ Redmine::Plugin.register :redmine_expert_metrics do
        { :controller => 'expert_metrics', :action => 'active_users' },
        menu_options
 end
+
+# Count delivered notification mails. Mail.register_observer ignores a module
+# that is already registered, so re-running init.rb (dev reload) is harmless.
+ActionMailer::Base.register_observer(RedmineExpertMetrics::MailObserver)